+++ Business/clsPerson.cs
@@ -26,18 +26,30 @@ namespace Portfolio.Business
 
         public async Task<long> addNewPerson(PersonDTO person)
         {
+            if (!clsValidation.IsValidPersonDTO(person))
+                return 0;
+
             var newId = await __Person.addNew(person);
             return newId;
         }
 
         public async Task<bool> updatePersonById(PersonDTO person)
         {
+            if (!clsValidation.IsValidPersonDTO(person))
+                return false;
+
+            if (!clsValidation.IsValidId(person.ID))
+                return false;
+
             var result = await __Person.updateById(person);
             return result;
         }
 
         public async Task<bool> deletePersonById(long ID)
         {
+            if (!clsValidation.IsValidId(ID))
+                return false;
+
             var result = await __Person.deleteById(ID);
             return result;
         }
