+++ Business/clsSocialLink.cs
@@ -31,18 +31,30 @@ namespace Portfolio.Business
 
         public async Task<long> addNewSocialLink(SocialLinkDTO socialLink)
         {
+            if (!clsValidation.IsValidSocialLinkDTO(socialLink))
+                return 0;
+
             var newId = await __SocialLink.addNew(socialLink);
             return newId;
         }
 
         public async Task<bool> updateSocialLinkById(SocialLinkDTO socialLink)
         {
+            if (!clsValidation.IsValidSocialLinkDTO(socialLink))
+                return false;
+
+            if (!clsValidation.IsValidId(socialLink.ID))
+                return false;
+
             var result = await __SocialLink.updateById(socialLink);
             return result;
         }
 
         public async Task<bool> deleteSocialLinkById(long ID)
         {
+            if (!clsValidation.IsValidId(ID))
+                return false;
+
             var result = await __SocialLink.deleteById(ID);
             return result;
         }
